refactor: Define package managers as a map
Defining the package managers as an array, with the list of package
manager names being hardcoded opened up the opportunity for either the
list of definitions or the name type to be updated without the other,
creating a bug.

By defining the package managers an an object, and deriving the name
type from the keys of that object, it is impossible for them to get
out of sync.